+++ cognite/powerops/resync/models/market/benchmark.py
@@ -7,7 +7,8 @@ from pydantic import Field
 
 from cognite.powerops.cdf_labels import AssetLabel
 from cognite.powerops.resync.models._base import NonAssetType
-from cognite.powerops.resync.models.market.base import Bid, Process
+
+from .base import Bid, Process, ShopTransformation
 
 
 class BenchmarkBid(Bid):
@@ -24,6 +25,7 @@ class BenchmarkProcess(Process):
     parent_description: ClassVar[str] = "Configurations used in benchmarking processes"
     description: str = "Configuration for benchmarking of day-ahead bidding"
     label: ClassVar[AssetLabel] = AssetLabel.DAYAHEAD_BIDDING_BENCHMARKING_CONFIG
+    shop: ShopTransformation
     bid: BenchmarkBid
     production_plan_time_series: list[ProductionPlanTimeSeries] = Field(default_factory=list)
     benchmarking_metrics: dict[str, str] = Field(default_factory=dict)